:alien:  Update

+++ CPP/serveurs/StructParc.h
@@ -13,12 +13,12 @@
 struct fich_parc
 {
     int id;
-    float x;
-    float y;
+    int x;
+    int y;
     int flagemplacement;
     char datereservation[MAX_CHAINE];
     char datearrivee[MAX_CHAINE];
-    float poids;
+    int poids;
     char destination[MAX_CHAINE];
     char moyenTransport[MAX_CHAINE];
 };
